+++ mistralrs-quant/src/utils/mod.rs
@@ -1,13 +1,16 @@
-use candle_core::{
-    cuda::{cudarc::driver::DevicePtr, CudaDType},
-    CudaDevice, Device, Storage, Tensor, WithDType,
-};
-
+#[cfg(feature = "cuda")]
 mod ffi;
 mod ops;
 
 pub use ops::{BitWiseOp, LeftshiftOp};
 
+#[cfg(feature = "cuda")]
+use candle_core::{
+    cuda::{cudarc::driver::DevicePtr, CudaDType},
+    CudaDevice, Device, Storage, Tensor, WithDType,
+};
+
+#[cfg(feature = "cuda")]
 pub(crate) fn get_cuda_slice<T: WithDType + CudaDType>(x: &Tensor) -> *const T {
     match &*x.storage_and_layout().0 {
         Storage::Cuda(a_storage) => *a_storage
@@ -18,6 +21,7 @@ pub(crate) fn get_cuda_slice<T: WithDType + CudaDType>(x: &Tensor) -> *const T {
     }
 }
 
+#[cfg(feature = "cuda")]
 pub(crate) fn get_cuda_device(x: &Tensor) -> &CudaDevice {
     match x.device() {
         Device::Cuda(dev) => dev,
